renamed ft_transfer to preparation_ft_transfer
