+++ src/commands/tokens/send_ft/amount_ft.rs
@@ -10,7 +10,7 @@ pub struct AmountFt {
     ft_transfer_amount: crate::types::ft_properties::FungibleTokenTransferAmount,
     #[interactive_clap(named_arg)]
     /// Enter gas for function call
-    prepaid_gas: super::ft_transfer::PrepaidGas,
+    prepaid_gas: super::preparation_ft_transfer::PrepaidGas,
 }
 
 #[derive(Debug, Clone)]
